% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_df_manipulation.R
\name{clean_names2}
\alias{clean_names2}
\title{Cleans the column-names for a dataset using `janitor::clean_names` and stores the mapping between
the old and new column-names as an attribute.}
\usage{
clean_names2(x, ...)
}
\arguments{
\item{x}{A data-frame.}

\item{...}{Additional arguments, passed onto `janitor::clean_names`.}
}
\value{
The same data-frame, but with the column names cleaned and an additional attribute
added. The attribute `colname_map` is a named vector containing a mapping between the old (the
names) and new (the values) column names in the data-frame. This vector is in the same structure
as expected by 'choices' in selectInput (the human-readable values are the names,
the computer-friendly strings are the values).
}
\description{
Cleans the column-names for a dataset using `janitor::clean_names` and stores the mapping between
the old and new column-names as an attribute.
}
