% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_app}
\alias{mod_table_app}
\title{Module for presenting / formatting a table}
\usage{
mod_table_app(id = "table", rx_plotting_dataset, rx_x_var, rx_y_var)
}
\description{
Module for presenting / formatting a table
}
