% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{un_data}
\alias{un_data}
\title{UN country statistics dataset (CC0)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 229 rows and 50 columns.
}
\usage{
un_data
}
\description{
Obtained from `https://www.kaggle.com/datasets/sudalairajkumar/undata-country-profiles`
on 2023-09-04.
}
\keyword{datasets}
